feat: Add rename_hosts mapping
This add rename_hosts mapping to providers config. This can become
useful for people using SSH to connect to Github/Gitlab and configured
custom hostname in their .ssh/config. For instance in SSH config Github
can be referenced as gh, and this can be used also in git's clone. This
will set push/pull urls for origin to gh:owner/repo, making it
impossible for pipeline.nvim to query APIs - it would try to make
request to https://gh/api/.... This change add a way to translate back
custom named provider hosts back to proper ones. Since this is for very
specific usecases, there is no point in automating it (and it works
without gh/glab cli commands)